[IJK/Algo_QC] Adding compatible Algo_QC 192, bug correction QUICK
Adding compilable Algo_QC with patched items that correspond to
the right calculation of surfaces for the calculation of
the mass flux in the QUICK and Centered order 2 schemes

Correction of the Simd tools used to compute the terms
of the QUICK scheme :

in the J direction (spanwise in a channel), the offset to take
into account in the J direction for each parallel domain was
fixed when it needed to be taken into account.
This had the effect of not computing the flux correctly on
the very first J cell for the mass flux.

This is a very similar bug to commit hash:
ead456d04d72824210683fa612a7885d41a20d1e

which is detailed in the changelogs of Algo_QC.